+++ src/components/Details/CastComponent.jsx
