+++ src/styles.scss
@@ -2,7 +2,7 @@
 @import './scss/globals.scss';
 @import './components/main/main.scss';
 @import "components/header/header";
-@import "components/FilmDetails/filmDetails";
+@import "components/Details/Details";
 @import "components/Divisor/Divisor";
